wow I didn't realize I did not commit this
Been running for two years with this FW version, and it turns out I had
not commited any of the changes, whoops.

Main change is displaying a second time zone. I was in US PNW and needed
to know Oslo time for calling the fam.

Looks like there's also refactoring of the date screen.
And BLE advertising is off by default.